requirements.txt
#List of required packages needing installation

Pillow==11.1.0  # Replace with the version you're using